requirements.txt
numpy~=2.3.0
protobuf~=5.29.5
sentence-transformers~=4.1.0
torch~=2.3.0
faiss-cpu~=1.11.0
pillow~=11.2.1
tqdm~=4.67.1
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.4
streamlit~=1.45.1